wip: event receivers controller
