+++ src/main/kotlin/it/pagopa/wallet/eventdispatcher/configuration/properties/RedisStreamEventControllerConfigs.kt
@@ -6,11 +6,8 @@ import org.springframework.boot.context.properties.ConfigurationProperties
 @ConfigurationProperties(prefix = "redis-stream.event-controller")
 class RedisStreamEventControllerConfigs(
     val streamKey: String,
-    consumerGroupPrefix: String,
     consumerNamePrefix: String,
-    val faiOnErrorCreatingConsumerGroup: Boolean
 ) {
     private val uniqueConsumerId = UUID.randomUUID().toString()
-    val consumerGroup = "$consumerGroupPrefix-$uniqueConsumerId"
     val consumerName = "$consumerNamePrefix-$uniqueConsumerId"
 }
