+++ src/main/kotlin/it/pagopa/wallet/eventdispatcher/service/InboundChannelAdapterLifecycleHandlerService.kt
@@ -1,6 +1,5 @@
 package it.pagopa.wallet.eventdispatcher.service
 
-import it.pagopa.wallet.eventdispatcher.configuration.redis.stream.RedisStreamMessageSource
 import it.pagopa.wallet.eventdispatcher.repositories.redis.bean.ReceiverStatus
 import it.pagopa.wallet.eventdispatcher.repositories.redis.bean.Status
 import org.springframework.beans.factory.annotation.Autowired
@@ -65,8 +64,5 @@ class InboundChannelAdapterLifecycleHandlerService(
 
     /** Retrieve all InboundChannelAdapter on which perform commands */
     fun findInboundChannelAdapterBeans() =
-        applicationContext
-            .getBeansWithAnnotation(InboundChannelAdapter::class.java)
-            .filterNot { it.value is RedisStreamMessageSource }
-            .keys
+        applicationContext.getBeansWithAnnotation(InboundChannelAdapter::class.java).keys
 }
