+++ src/main/resources/application.properties
@@ -16,7 +16,7 @@ server.servlet.context-path=/
 server.port=8080
 # Logging
 logging.level.root=\${DEFAULT_LOGGING_LEVEL:INFO}
-logging.level.it.gov.pagopa=\${APP_LOGGING_LEVEL:INFO}
+logging.level.it.pagopa=\${APP_LOGGING_LEVEL:INFO}
 # Azure Storage Queue
 spring.cloud.azure.storage.queue.enable=true
 spring.cloud.azure.storage.queue.account-key=\${WALLET_STORAGE_QUEUE_KEY}
@@ -49,8 +49,6 @@ spring.redis.ssl=\${REDIS_SSL_ENABLED}
 # Redis stream configuration
 redis-stream.event-controller.streamKey=\${REDIS_STREAM_EVENT_CONTROLLER_STREAM_KEY}
 redis-stream.event-controller.consumerGroupPrefix=\${REDIS_STREAM_EVENT_CONTROLLER_CONSUMER_GROUP_PREFIX}
-redis-stream.event-controller.consumerNamePrefix=\${REDIS_STREAM_EVENT_CONTROLLER_CONSUMER_NAME_PREFIX}
-redis-stream.event-controller.faiOnErrorCreatingConsumerGroup=\${REDIS_STREAM_EVENT_CONTROLLER_FAIL_ON_ERROR_CREATING_CONSUMER_GROUP:true}
 # event receivers controller configurations
 eventController.status.pollingChron=\${EVENT_CONTROLLER_STATUS_POLLING_CHRON}
 eventController.deploymentVersion=\${DEPLOYMENT_VERSION:PROD}
